refactor(triage): split github_triage.py (929 LOC) into a 7-file package
`shared/scripts/github_triage.py` is deleted and replaced by the package
directory `shared/scripts/github_triage/`:

- `__init__.py`  — re-exports the public surface (12 names + SOURCE).
- `state.py`     — throttle state (is_due, read/write_last_import,
                   throttle_hours, DEFAULT_THROTTLE_HOURS).
- `severity.py`  — severity vocab + per-feed extractors
                   (triage_severity + internal _cs/_db/_artifact extractors,
                   url helpers, kind/max/breakdown helpers).
- `producer.py`  — security action-unit mappers + owned prefix constants
                   (security_action_unit, security_action_unit_from_artifact,
                   PREFIX_SECURITY/SECRETS/CI).
- `mappers.py`   — secrets + CI mappers (secrets_action_unit,
                   ci_action_unit, latest_failed_ci_runs).
- `resolve.py`   — auto-resolve + legacy-migration sweeps (resolve_stale,
                   migrate_legacy_items, SOURCE).
- `consumer.py`  — import_findings orchestrator (the side-effectful entry).

Every submodule is <= 300 LOC.

The public import surface is preserved exactly: `import github_triage`
keeps working for the SessionStart hook
(`shared/scripts/hooks/import_github_findings.py`) and for all three test
suites (`shared/tests/test_github_triage*.py`), which patch
`github_triage.is_due`, `github_triage.import_findings`, and
`github_triage.write_last_import` via monkeypatch — same module-object
semantics as pre-split.

Empirical verification:
- 72/73 tests pass across the three github_triage suites. The remaining
  failure (`test_failed_fetch_does_not_resolve_items`) is pre-existing on
  main pre-split and out of scope for B6.
- Round-trip probes: import-surface resolution, hook-caller import flow,
  monkeypatch-through-hook, and producer→consumer end-to-end all pass.

CLEANUP-INVARIANT (campaign B6):
- `shared/scripts/github_triage.py` deleted -> baseline entry REMOVED
  per rule (b) in the same atomic commit.
- No new baseline entries; all submodules are <= 300 LOC.

Sub-iterate of campaign 2026-05-25-bloat-cleanup-B-shipwright.
