+++ CHANGELOG-unreleased.d/Changed/iterate-2026-05-25-bloat-B-github-triage_001.md
@@ -1,0 +1,1 @@
+- (refactor) Split `shared/scripts/github_triage.py` from 929 LOC into a 7-file package (`shared/scripts/github_triage/`): `__init__.py` (re-exports the public surface), `state.py` (throttle state), `severity.py` (severity vocab + per-feed extractors), `producer.py` (security action-unit mappers), `mappers.py` (secrets + CI action-unit mappers), `resolve.py` (auto-resolve + legacy-migration sweeps), `consumer.py` (`import_findings` orchestrator). Each submodule <= 300 LOC; public import surface (`import_findings`, `is_due`, `read_last_import`, `write_last_import`, `throttle_hours`, `DEFAULT_THROTTLE_HOURS`, `triage_severity`, `latest_failed_ci_runs`, `security_action_unit`, `security_action_unit_from_artifact`, `secrets_action_unit`, `ci_action_unit`) preserved unchanged — `shared/scripts/hooks/import_github_findings.py` and `shared/tests/test_github_triage*.py` keep working. `shipwright_bloat_baseline.json` entry removed per cleanup invariant rule (b). Campaign B sub-iterate B6.
