+++ shipwright_bloat_baseline.json
@@ -541,13 +541,6 @@
       "adr": null
     },
     {
-      "path": "shared/scripts/github_triage.py",
-      "limit": 300,
-      "current": 929,
-      "state": "grandfathered",
-      "adr": null
-    },
-    {
       "path": "shared/scripts/hooks/audit_phase_quality_on_stop.py",
       "limit": 300,
       "current": 310,
